// pavi-front/src/app/core/models/models.ts
export const POSICAO_FROTA_STATUS = [
  'Aguardando Descarga',
  'Aguardando Descarga na Diária',
  'Aguardando Carregar',
  'Aguardando Carregar na Diária',
  'Aguardando Troca de Nota',
  'Carregando',
  'Descarregando',
  'Desloc Carregar',
  'Carregado/Folga',
  'Vazio',
  'Viajando',
  'Vazio/Aguardando Canhoto',
  'Aguardando NF',
  'Repaletizando',
  'Folga',
] as const;

export type StatusType = typeof POSICAO_FROTA_STATUS[number];
// ...
export interface StatusMeta {
  solid: string;
  fg: string;
  bg: string;
  border: string;
  label: string;
}